feat: detail related videos
