+++ src/components/detail/related-videos/RelatedVideos.module.css
@@ -1,0 +1,26 @@
+.related {
+    display: flex;
+    flex-direction: column;
+
+    width: 100%;
+    height: 100vh;
+
+    gap: 0.375rem;
+}
+
+.item {
+    display: flex;
+
+    gap: 1.25rem;
+}
+
+.video {
+    flex: 1 1 auto;
+    height: 6rem;
+
+    background-color: lightsalmon;
+}
+
+.content {
+    flex: 1 1 auto;
+}
