+++ src/components/layout/Layout.module.css
@@ -2,5 +2,5 @@
     display: flex;
     flex-direction: column;
 
-    width: 90%;
+    width: 74%;
 }
