+++ src/components/layout/header/SearchHeader.jsx
@@ -1,17 +1,39 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { BiSearch } from 'react-icons/bi';
+import { useNavigate } from 'react-router-dom';
 import styles from './SearchHeader.module.css';
 
-export default function SearchHeader({ handleSubmit, handleChangeKeyword, keyword }) {
+export default function SearchHeader() {
+    const navigate = useNavigate();
+    const [keyword, setKeyword] = useState('');
+
+    const handleChangeKeyword = (e) => {
+        setKeyword(e.target.value);
+    };
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+
+        navigate(`/${keyword}`);
+
+        // 새로 검색하고 필터한다
+    };
+
     return (
         <header>
-            <div className={styles.logo}>
+            <div onClick={() => navigate('/')} className={styles.logo}>
                 {/* <img src='assets/logo.png' alt='logo' /> */}
                 <img src='assets/logo-dark.png' alt='logo' />
             </div>
 
             <form onSubmit={handleSubmit}>
-                <input className={styles.input} onChange={handleChangeKeyword} type='text' value={keyword} />
+                <input
+                    className={styles.input}
+                    onChange={handleChangeKeyword}
+                    type='text'
+                    value={keyword}
+                    placeholder='Search...'
+                />
                 <button type='button'>
                     <BiSearch />
                 </button>
