+++ src/components/layout/header/SearchHeader.module.css
@@ -8,10 +8,8 @@ header {
 
     border-bottom: 1px solid #bebebe;
 
-    padding: 1.25rem 0;
+    padding: 1rem 0;
     padding-right: 10rem;
-
-    margin-bottom: 2.5rem;
 }
 
 .logo {
