+++ src/pages/DetailPage.jsx
@@ -2,6 +2,7 @@ import Content from 'components/detail/content/Content';
 import RelatedVideos from 'components/detail/related-videos/RelatedVideos';
 import React from 'react';
 import { useParams } from 'react-router-dom';
+import styles from './Pages.module.css';
 
 export default function DetailPage() {
     const { videoId } = useParams();
@@ -9,7 +10,7 @@ export default function DetailPage() {
     console.log(videoId);
 
     return (
-        <div>
+        <div className={styles.detail}>
             <Content />
             <RelatedVideos />
         </div>
