+++ src/pages/Pages.module.css
@@ -1,6 +1,11 @@
-.page {
-    display: flex;
-    flex-direction: column;
+.main {
+    margin-top: 40px;
+}
 
-    width: 100%;
+.detail {
+    display: flex;
+    justify-content: space-between;
+    gap: 10px;
+
+    margin-top: 1rem;
 }
